+++ app/controllers/crono/jobs_controller.rb
@@ -1,11 +1,15 @@
 module Crono
   class JobsController < ApplicationController
     def index
-      @jobs = Crono::CronoJob.all
+      render json: {
+        failed: Crono::CronoJob.by_failed,
+        succeed: Crono::CronoJob.by_success,
+        new: Crono::CronoJob.by_new
+      }, status: :ok
     end
 
     def show
-      @job = Crono::CronoJob.find(params[:id])
+      render json: Crono::CronoJob.find(params[:id]), status: :ok
     end
   end
 end