+++ config/routes.rb
@@ -1,4 +1,0 @@
-Crono::Engine.routes.draw do
-  resources :jobs
-  root 'jobs#index'
-end