+++ lib/crono/period.rb
@@ -1,8 +1,7 @@
 module Crono
   # Period describe frequency of jobs
   class Period
-    DAYS = [:monday, :tuesday, :wednesday, :thursday, :friday, :saturday,
-            :sunday]
+    DAYS = %i[monday tuesday wednesday thursday friday saturday sunday]
 
     def initialize(period, at: nil, on: nil, within: nil)
       @period = period